Move resx files in NTS project
https://github.com/Not-Endurance/not-timing-system/issues/409
