+++ src/Judge/NTS.Judge.Blazor/GlobalUsings.cs
@@ -1,4 +1,4 @@
 ﻿global using Microsoft.AspNetCore.Components;
 global using NTS.Domain.Enums;
-global using static Not.Localization.NLocalizedStrings;
-global using static NTS.Localization.NtsLocalizedStrings;
+global using static Not.Localization.NStrings;
+global using static NTS.Localization.NtsStrings;
