+++ src/Judge/NTS.Judge.Blazor/Nexus/InquiryPage.razor
@@ -1,7 +1,7 @@
 ﻿@attribute [Route(INQUIRY_PAGE)]
 @inherits NComponent
 
-<NContent Title="Inquiry performances">
+<NContent Title="@Inquiry_string">
     
     <Main>
         <NSelect T="InquiryType?" @bind-Value="Type" />
