--- src/Judge/NTS.Judge.Blazor/Resources/Localization/Strings.cs
@@ -1,7 +1,0 @@
-﻿namespace NTS.Judge.Blazor.Resources.Localization;
-
-// This class is used Type parameter in IStringLocalizer<T>
-// to serve as a convetional anchor to locate the resx file
-public class Strings 
-{ 
-}
