+++ src/Judge/NTS.Judge.Blazor/_Imports.razor
@@ -9,11 +9,10 @@
 @using Not.Blazor
 @using Not.Blazor.Components
 @using Not.Blazor.Navigation;
-@using static Not.Localization.NLocalizedStrings;
-@using static NTS.Localization.NtsLocalizedStrings;
+@using static Not.Localization.NStrings;
+@using static NTS.Localization.NtsStrings;
 @using NTS.Judge.Blazor
 @using NTS.Judge.Blazor.Shared
-@using NTS.Judge.Blazor.Resources.Localization
 @using NTS.Domain
 @using NTS.Domain.Enums;
 @using MudBlazor
