+++ src/NTS.Domain.Core/GlobalUsings.cs
@@ -3,5 +3,5 @@ global using Not.Exceptions;
 global using Not.Extensions;
 global using NTS.Domain.Enums;
 global using NTS.Domain.Objects;
-global using static Not.Localization.NLocalizedStrings;
-global using static NTS.Localization.NtsLocalizedStrings;
+global using static Not.Localization.NStrings;
+global using static NTS.Localization.NtsStrings;
