+++ src/NTS.Domain.Setup/NTS.Domain.Setup.csproj
@@ -8,10 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\NTS.Domain\NTS.Domain.csproj" />
     <ProjectReference Include="..\Not\Not.Domain\Not.Domain.csproj" />
     <ProjectReference Include="..\Not\Not\Not.csproj" />
