+++ src/NTS.Domain/GlobalUsings.cs
@@ -6,5 +6,5 @@ global using System.Text;
 global using System.Threading.Tasks;
 global using Not.Domain;
 global using NTS.Domain.Objects;
-global using static Not.Localization.NLocalizedStrings;
-global using static NTS.Localization.NtsLocalizedStrings;
+global using static Not.Localization.NStrings;
+global using static NTS.Localization.NtsStrings;
