+++ src/NTS/Localization/NtsStrings.cs
@@ -7,7 +7,7 @@
 
 namespace NTS.Localization
 {
-    public static class NtsLocalizedStrings
+    public static class NtsStrings
     {
         public static string km_string => LocalizeString(nameof(km_string));
         public static string rest_string => LocalizeString(nameof(rest_string));
