+++ src/NTS/Localization/Resources/LocalizedStrings.bg.resx
@@ -129,7 +129,7 @@
     <value>Разпечатки</value>
   </data>
   <data name="Inquiry_string" xml:space="preserve">
-    <value>Запитване</value>
+    <value>Справки</value>
   </data>
   <data name="Failed_to_Qualify_Reason_string" xml:space="preserve">
     <value>Причина за отпадане</value>
