+++ src/NTS/Localization/Resources/LocalizedStrings.cs
@@ -1,0 +1,7 @@
+﻿namespace NTS.Localization.Resources;
+
+// This class is used Type parameter in IStringLocalizer<T>
+// and converted to a resource during build
+public class LocalizedStrings 
+{ 
+}
