+++ src/NTS/Localization/Resources/LocalizedStrings.resx
