+++ src/Nexus/NTS.Nexus.HTTP/GlobalUsings.cs
@@ -1,2 +1,2 @@
-﻿global using static Not.Localization.NLocalizedStrings;
-global using static NTS.Localization.NtsLocalizedStrings;
+﻿global using static Not.Localization.NStrings;
+global using static NTS.Localization.NtsStrings;
