+++ src/Not/Not.Blazor/GlobalUsings.cs
@@ -1,4 +1,4 @@
 ﻿global using Microsoft.AspNetCore.Components;
 global using Not.Exceptions;
-global using static Not.Localization.NLocalizedStrings;
+global using static Not.Localization.NStrings;
