+++ src/Not/Not.Blazor/_Imports.razor
@@ -5,6 +5,6 @@
 @using Not.Localization
 @using Not.Domain
 @using Not.Blazor.Components
-@using static Not.Localization.NLocalizedStrings;
+@using static Not.Localization.NStrings;
 
 @using MudBlazor
