+++ src/Not/Not.Storage/GlobalUsings.cs
@@ -1,3 +1,3 @@
 ﻿global using Not.Domain;
 global using Not.Storage.Stores;
-global using static Not.Localization.NLocalizedStrings;
+global using static Not.Localization.NStrings;
